Not working.  Started setting up server and package.json
Not happy with hacky way I have it building the schema (it is not actually building right now even though it says it has).  Had issue with using imports, and still do inside the graphql.js file. Not sure why

Need to review server/index.js and possibly pull out graphql content and put in middleware?

Need to edit the npm start command to work in graphql building its schema and something to do with build file